backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
google-generativeai
google-genai
